refactor(yggd): move worker functions to worker.go
